Major changes

+++ main.py
@@ -1,27 +1,31 @@
-import time
-import asyncio
 import traceback
+import asyncio
+import time
+import sys
 
-from discord import Client
-from discord import Game
+from aiohttp import ClientSession
 
-from commandmanager import CommandManager
+import discord
+
+from modulemanager import ModuleManager
 
 from utils.constants import PREFIXES
 from utils.formatters import format_response
 from utils.config import Config
 
 
-class BotMyBot(Client):
+class BotMyBot(discord.Client):
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, start_time, *args, **kwargs):
         super(BotMyBot, self).__init__(*args, **kwargs)
+        self.start_time = start_time
+        self.session = ClientSession()
+
         self.config = Config('config.json', loop=self.loop)
         self.token = self.config['token']
-        self.cm = CommandManager(self)
-        self.cm.load_commands()
 
         self.tracked_messages = {}
+        self.mm = ModuleManager(self)
 
     def run(self, token=None):
         if token is None:
@@ -29,14 +33,23 @@ class BotMyBot(Client):
         else:
             self.token = token
 
-        super(BotMyBot, self).run(token)
+        super(BotMyBot, self).run(token, reconnect=True)
+
+    async def restart(self):
+        await self.stop(0)
+
+    async def stop(self, exit_code=0):
+        await self.close()
+        sys.exit(exit_code)
 
     async def on_ready(self):
+        await self.mm.load_modules()
+        print('Loaded modules: [%s]' % ' '.join(self.mm.modules.keys()))
         print('Ready')
 
-        previous_status = self.config.get('previous_status', '')
-        if previous_status:
-            await self.change_presence(game=Game(name=previous_status))
+    async def close(self):
+        await super(BotMyBot, self).close()
+        print('Closed')
 
     async def on_message(self, message, from_edit=False):
         if message.author.bot:
@@ -55,19 +68,25 @@ class BotMyBot(Client):
             return
 
         message.content = message.content[
-            len(next(p for p in PREFIXES if lower_content.startswith(p))):]
+            len(next(p for p in PREFIXES if lower_content.startswith(p))):].strip()
+        
+        if not message.content:
+            return
 
-        command_response = await self.cm.check_commands(message)
+        module_response = await self.mm.check_modules(message)
 
-        if command_response:
-            command_response = await format_response(
-                command_response, message, self)
+        if module_response:
+            module_response = await format_response(
+                module_response, message, self)
 
-        if command_response:
+        if module_response:
             await self.send_message(
-                message.channel, command_response, response_to=message)
+                message.channel, module_response, response_to=message)
 
     async def on_message_edit(self, before, after):
+        if before.content == after.content:
+            return
+
         if before.id in self.tracked_messages:
             if self.tracked_messages[before.id] is not None:
                 await self.delete_message(self.tracked_messages[before.id])
@@ -82,10 +101,14 @@ class BotMyBot(Client):
                 self.tracked_messages[message.id] = None
 
 
-    async def send_message(self, destination, *args, response_to=None, **kwargs):
+    async def send_message(self, destination, text, *args, response_to=None, **kwargs):
+        text = text.replace(self.token, 'my-token')
+        text = text.replace('@everyone', '@\u200beveryone')
+        text = text.replace('@here', '@\u200bhere')
+
         try:
             message = await super(BotMyBot, self).send_message(
-                destination, *args, **kwargs)
+                destination, text, *args, **kwargs)
         except Exception:
             exception = traceback.format_exc()
             exception = '\n'.join(exception.split('\n')[-4:])
@@ -97,6 +120,31 @@ class BotMyBot(Client):
                 await self.register_response(response_to, message)
 
             return message
+
+    async def edit_message(self, message, text, *args, **kwargs):
+        message.content = message.content.replace(self.token, 'my-token')
+        message.content = message.content.replace('@everyone', '@\u200beveryone')
+        message.content = message.content.replace('@here', '@\u200bhere')
+
+        try:
+            return await super(BotMyBot, self).edit_message(
+                message, text, *args, **kwargs)
+        except discord.errors.NotFound:
+            print('edit_message: message not found')
+            return
+        except Exception:
+            exception = traceback.format_exc()
+            exception = '\n'.join(exception.split('\n')[-4:])
+            exception = '❗ Message edit failed\n```\n' + exception + '```'
+            return await super(BotMyBot, self).edit_message(
+                message, exception, *args, **kwargs)
+    
+    async def delete_message(self, *args, **kwargs):
+        try:
+            return await super(BotMyBot, self).delete_message(*args, **kwargs)
+        except discord.errors.NotFound:
+            print('delete_message: message not found')
+            return
 
     async def track_message(self, message):
         self.tracked_messages[message.id] = None
@@ -113,4 +161,4 @@ class BotMyBot(Client):
 
 
 if __name__ == '__main__':
-    BotMyBot().run()
+    BotMyBot(time.time()).run()
